fix error

+++ src/BirdStoreAndroid/app/src/main/java/com/example/birdstoreandroid/Model/GetCartResponse.java
@@ -36,7 +36,7 @@ public class GetCartResponse {
         private String id;
         private String product_id;
         private String order_id;
-        private int price;
+        private float price;
         private int quantity;
         private String user_id;
         private Product product;
@@ -68,11 +68,11 @@ public class GetCartResponse {
             this.order_id = order_id;
         }
 
-        public int getPrice() {
+        public float getPrice() {
             return price;
         }
 
-        public void setPrice(int price) {
+        public void setPrice(float price) {
             this.price = price;
         }
 
@@ -129,7 +129,7 @@ public class GetCartResponse {
         private String id;
         private String category_id;
         private String image;
-        private int price;
+        private float price;
         private String name;
         private boolean sex;
         private String description;
@@ -161,11 +161,11 @@ public class GetCartResponse {
             this.image = image;
         }
 
-        public int getPrice() {
+        public float getPrice() {
             return price;
         }
 
-        public void setPrice(int price) {
+        public void setPrice(float price) {
             this.price = price;
         }
